chore(lint): sweep lint findings on the vec tracing changeset
Follow-up to ad4638b7 to clear make lint:

  - gci: fix import grouping in api/data/codec_test.go and
    banyand/query/tracing_test.go (apache imports were split across two
    blocks; third-party google.golang.org/protobuf belongs in the
    second block, apache prefix in the third).
  - gci: drop the spurious blank line between two helpers in
    pkg/query/vectorized/measure/plan/distributed.go.
  - govet/fieldalignment: reorder BatchAggregation, BatchLimit, and
    rawFrameFuture struct fields by descending size to minimise padding.
  - govet/exhaustive: list AggModeAll explicitly in the two switches
    over a.mode so the default-case fallthrough is no longer implicit.
  - gocritic/stringXbytes: use bytes.Equal instead of string conversion
    in tracing_collect_test.go and tracing_rolling_upgrade_test.go.
  - gocritic/appendAssign: build the merged future slice with explicit
    allocation in tracing_rolling_upgrade_test.go.
  - gocritic/commentedOutCode: rephrase three comments whose ASCII math
    triggered the "commented-out code" heuristic.
  - misspell: normalise British → American (behaviour, summarise,
    prioritised, summarised) in tracing_summary.go and the matching test.
  - unparam: drop the minFactor argument from assertSpanIsBottleneck;
    the threshold is now a package-level const bottleneckMinFactor.
  - unconvert: drop the redundant int64() wrap around rows*8.

No behaviour change. make lint and the tracing-related package tests
remain green.

via [HAPI](https://hapi.run)

+++ pkg/query/vectorized/measure/plan/tracing_summary_test.go
@@ -142,7 +142,7 @@ func assertTagInt64Ordered(t *testing.T, span *commonv1.Span, keys ...string) {
 //   - 5 zero-row nodes (nodeIDs 4–8, latency 80 ms each)
 //   - 17 healthy nodes (nodeIDs 9–25, latencies 10–170 ms, 10 ms steps)
 //
-// Expected behaviour:
+// Expected behavior:
 //   - Exactly 19 individual children on the broadcast span + 1 data-summary child.
 //   - The 19 individual slots are filled by: 3 error + 5 zero-row + 11 highest-latency healthy nodes.
 //   - data-summary.aggregated_data_node_spans = 6  (the remaining 6 lower-latency healthy nodes).
@@ -167,7 +167,7 @@ func TestFanoutDataSummary_AtThreshold(t *testing.T) {
 	for idx := range 17 {
 		latencyMS := int64((17 - idx) * 10) // 170, 160, ..., 10
 		rows := int64((17 - idx) * 100)
-		nodes = append(nodes, makeNodeInfo(idx+9, latencyMS*int64(time.Millisecond), rows, int64(rows*8), false))
+		nodes = append(nodes, makeNodeInfo(idx+9, latencyMS*int64(time.Millisecond), rows, rows*8, false))
 	}
 
 	applyFanoutCap(broadcastSpanCtx, broadcastSpan, nodes)
@@ -204,7 +204,7 @@ func TestFanoutDataSummary_AtThreshold(t *testing.T) {
 	assertTagPresent(t, lastChild, tracelabels.TagNodeLatencyNSMin)
 	assertTagPresent(t, lastChild, tracelabels.TagNodeLatencyNSMax)
 
-	// The 6 summarised nodes are all healthy (the 6 lowest-latency healthy ones: 10–60 ms).
+	// The 6 summarized nodes are all healthy (the 6 lowest-latency healthy ones: 10–60 ms).
 	// So nodes_with_errors = 0 and nodes_with_zero_rows = 0 in the summary.
 	assertTagValueEquals(t, lastChild, tracelabels.TagNodesWithErrors, "0")
 	assertTagValueEquals(t, lastChild, tracelabels.TagNodesWithZeroRows, "0")
@@ -289,11 +289,11 @@ func TestDecodeFrameSummary_AtThreshold(t *testing.T) {
 		t.Fatalf("child message = %q, want decode-frame-summary", summarySpan.GetMessage())
 	}
 
-	// frames_total = 100.
+	// Every frame is reflected in frames_total; no individual frame spans
+	// are emitted so frames_emitted_individually is 0 and frames_skipped
+	// equals the total.
 	assertTagValueEquals(t, summarySpan, tracelabels.TagFramesTotal, "100")
-	// frames_emitted_individually = 0 (no individual frame spans).
 	assertTagValueEquals(t, summarySpan, tracelabels.TagFramesEmittedIndividually, "0")
-	// frames_skipped = 100.
 	assertTagValueEquals(t, summarySpan, tracelabels.TagFramesSkipped, "100")
 
 	// All decode NS tags present.
